finalized logger config
