--- next-logger.config.cjs
@@ -1,3 +1,0 @@
-const logger = require("./logger/logger.cjs");
-
-module.exports = { logger };
