+++ package.json
@@ -35,9 +35,7 @@
     "mantine-datatable": "^2.9.12",
     "next": "^13.4.12",
     "next-auth": "^4.23.1",
-    "next-logger": "^3.0.2",
     "pino": "^8.15.0",
-    "pino-caller": "^3.4.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "superjson": "1.13.1",
